+++ service-destiny/src/routes/network.rs
@@ -41,13 +41,13 @@ async fn network_breakdown_season(
     let modes = if let Some(input_modes) = report_queries.modes {
         input_modes
             .split(',')
-            .map(|v| v.parse::<i32>().unwrap_or_default())
-            .collect::<Vec<i32>>()
+            .map(|v| v.parse::<i64>().unwrap_or_default())
+            .collect::<Vec<i64>>()
     } else {
         Vec::new()
     };
 
-    let season_input_number = season.parse::<i32>().unwrap_or_default();
+    let season_input_number = season.parse::<i64>().unwrap_or_default();
 
     let season = database::seasons::get(season_input_number, &state.database).await;
     let (season_start, season_end, season_number) = match season {
@@ -79,8 +79,8 @@ async fn network_breakdown_lifetime(
     let modes = if let Some(input_modes) = report_queries.modes {
         input_modes
             .split(',')
-            .map(|v| v.parse::<i32>().unwrap_or_default())
-            .collect::<Vec<i32>>()
+            .map(|v| v.parse::<i64>().unwrap_or_default())
+            .collect::<Vec<i64>>()
     } else {
         Vec::new()
     };
@@ -90,7 +90,7 @@ async fn network_breakdown_lifetime(
         .unwrap_or_default();
     // fetch from db
     let network_breakdown =
-        app::clan::network_breakdown(&modes, destiny2_launch_month_start.timestamp() as u64, 0, &mut state).await;
+        app::clan::network_breakdown(&modes, destiny2_launch_month_start.timestamp(), 0, &mut state).await;
 
     // convert to our response type
     let network_breakdown = network_breakdown
@@ -113,8 +113,8 @@ async fn network_lifetime_report(
     let modes = if let Some(input_modes) = report_queries.modes {
         input_modes
             .split(',')
-            .map(|v| v.parse::<i32>().unwrap_or_default())
-            .collect::<Vec<i32>>()
+            .map(|v| v.parse::<i64>().unwrap_or_default())
+            .collect::<Vec<i64>>()
     } else {
         Vec::new()
     };
@@ -153,8 +153,8 @@ async fn network_season_report(
     let modes = if let Some(input_modes) = report_queries.modes {
         input_modes
             .split(',')
-            .map(|v| v.parse::<i32>().unwrap_or_default())
-            .collect::<Vec<i32>>()
+            .map(|v| v.parse::<i64>().unwrap_or_default())
+            .collect::<Vec<i64>>()
     } else {
         Vec::new()
     };
@@ -165,7 +165,7 @@ async fn network_season_report(
         let (task_started, report) = app::report::member::season(
             member.membership_id.to_string(),
             &modes,
-            season.parse::<i32>().unwrap_or_default(),
+            season.parse::<i64>().unwrap_or_default(),
             true,
             &mut state,
         )
